+++ aivm/cli/vm.py
@@ -18,6 +18,7 @@ from .vm_attach import (
 )
 from .vm_config import VMEditCLI
 from .vm_cache import VMFlushCachesCLI
+from .vm_guard import VMFdGuardCLI
 from .vm_connect import VMCodeCLI, VMSSHCLI, VMSshConfigCLI, VMWaitIPCLI
 from .vm_lifecycle import (
     VMCreateCLI,
@@ -56,6 +57,7 @@ class VMModalCLI(kwconf.ModalCLI):
     )
     code = VMCodeCLI
     flush_caches = VMFlushCachesCLI
+    fdguard = VMFdGuardCLI
 
 
 __all__ = [
@@ -79,4 +81,5 @@ __all__ = [
     'VMEditCLI',
     'VMUpdateCLI',
     'VMFlushCachesCLI',
+    'VMFdGuardCLI',
 ]